LPS-136999 Move RouterImpl.java to a public package for reuse
