+++ portal-impl/src/com/liferay/portlet/internal/FriendlyURLMapperTrackerImpl.java
@@ -33,6 +33,7 @@ import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.xml.Document;
 import com.liferay.portal.kernel.xml.Element;
 import com.liferay.portal.kernel.xml.UnsecureSAXReaderUtil;
+import com.liferay.portlet.RouterImpl;
 
 import java.util.List;
 import java.util.Map;
