+++ portal-impl/src/com/liferay/portlet/packageinfo
@@ -1,1 +1,1 @@
-version 4.0.0
+version 4.1.0
