+++ portal-impl/test/unit/com/liferay/portlet/RouteImplTest.java
@@ -19,7 +19,6 @@ import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.HttpUtil;
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
 import com.liferay.portal.util.HttpImpl;
-import com.liferay.portlet.internal.RouteImpl;
 
 import java.util.HashMap;
 import java.util.Map;
